Request-Response models moved to seperate ptoject
