+++ SoftwareEngineerInterview/CSharp/Zip.InstallmentsService/Zip.EmiCalc.RequestResponseModels/Zip.EmiCalc.RequestResponseModels.csproj
@@ -1,0 +1,9 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+  </PropertyGroup>
+
+</Project>
