+++ SoftwareEngineerInterview/CSharp/Zip.InstallmentsService/Zip.InstallmentsService.sln
@@ -6,11 +6,13 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Zip.InstallmentsService", "
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Zip.InstallmentsService.Test", "Zip.InstallmentsService.Test\Zip.InstallmentsService.Test.csproj", "{DF298FE5-362D-44BD-AD57-7DFC04CEB8B7}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Zip.EmiCalc.Api", "Zip.EmiCalc.Api\Zip.EmiCalc.Api.csproj", "{B306FBE0-27EF-40D6-AD37-C630A97E4687}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Zip.EmiCalc.Api", "Zip.EmiCalc.Api\Zip.EmiCalc.Api.csproj", "{B306FBE0-27EF-40D6-AD37-C630A97E4687}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Zip.EmiCalc.BusinessLogic", "Zip.EmiCalc.BusinessLogic\Zip.EmiCalc.BusinessLogic.csproj", "{8F29F1E6-052F-4869-800E-960C5D5CE196}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Zip.EmiCalc.BusinessLogic", "Zip.EmiCalc.BusinessLogic\Zip.EmiCalc.BusinessLogic.csproj", "{8F29F1E6-052F-4869-800E-960C5D5CE196}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Zip.EmiCalc.DataAccessRepository", "Zip.EmiCalc.DataAccessRepository\Zip.EmiCalc.DataAccessRepository.csproj", "{E79028FC-D802-4E7D-85B0-D02736F21DD9}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Zip.EmiCalc.DataAccessRepository", "Zip.EmiCalc.DataAccessRepository\Zip.EmiCalc.DataAccessRepository.csproj", "{E79028FC-D802-4E7D-85B0-D02736F21DD9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Zip.EmiCalc.RequestResponseModels", "Zip.EmiCalc.RequestResponseModels\Zip.EmiCalc.RequestResponseModels.csproj", "{0F700671-A7A5-45D4-9F49-373B3D65892C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -82,6 +84,18 @@ Global
 		{E79028FC-D802-4E7D-85B0-D02736F21DD9}.Release|x64.Build.0 = Release|Any CPU
 		{E79028FC-D802-4E7D-85B0-D02736F21DD9}.Release|x86.ActiveCfg = Release|Any CPU
 		{E79028FC-D802-4E7D-85B0-D02736F21DD9}.Release|x86.Build.0 = Release|Any CPU
+		{0F700671-A7A5-45D4-9F49-373B3D65892C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0F700671-A7A5-45D4-9F49-373B3D65892C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0F700671-A7A5-45D4-9F49-373B3D65892C}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{0F700671-A7A5-45D4-9F49-373B3D65892C}.Debug|x64.Build.0 = Debug|Any CPU
+		{0F700671-A7A5-45D4-9F49-373B3D65892C}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{0F700671-A7A5-45D4-9F49-373B3D65892C}.Debug|x86.Build.0 = Debug|Any CPU
+		{0F700671-A7A5-45D4-9F49-373B3D65892C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0F700671-A7A5-45D4-9F49-373B3D65892C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{0F700671-A7A5-45D4-9F49-373B3D65892C}.Release|x64.ActiveCfg = Release|Any CPU
+		{0F700671-A7A5-45D4-9F49-373B3D65892C}.Release|x64.Build.0 = Release|Any CPU
+		{0F700671-A7A5-45D4-9F49-373B3D65892C}.Release|x86.ActiveCfg = Release|Any CPU
+		{0F700671-A7A5-45D4-9F49-373B3D65892C}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
